% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar_fun}
\alias{bar_fun}
\title{Bar plot}
\usage{
bar_fun(df, year, index, path = "output/EDA/plots/", na.rm = T, ...)
}
\arguments{
\item{df1, year, index, path}{Input dataframe, year, and path is default to current}
}
\value{
bar plots
}
\description{
Bar plot
}
\details{
This function creates an output of bar plots.
}
