% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(df, beg_row = 1, end_row = 2, lower_bound = 3, upper_bound = 95)
}
\arguments{
\item{df, beg_row, end_row, lower_bound, upper_bound}{Input dataframe, start row defaults to 1 and end row defaults to 2, lower bound defaults to 3 and upper to 95}
}
\value{
cleaned dataframe
}
\description{
Step 2
}
\details{
This function removes the first two rows of the dataset which has name and
description information for each variable. All variable names are
standardized to lowercase and separated by underscores using the janitor
package dependency. Repeated IDs are removed.
}
