% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{completion_status}
\alias{completion_status}
\title{Separate files by completion status}
\usage{
completion_status(df, lower_bound = 3, upper_bound = 95)
}
\arguments{
\item{df, lower_bound, upper_bound}{Input dataframe, lower_bound defaults to 3, upper_bound defaults to 95}
}
\value{
df with additional status variable
}
\description{
Step 6
}
\details{
This function adds a completion status column based on the progress variable.
The lower and upper bounds on what is considered complete can be manipulated.
}
