% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{create_codebook}
\alias{create_codebook}
\title{Create codebook/ key file}
\usage{
create_codebook(df, year, path = "")
}
\arguments{
\item{df, year, path}{Input dataframe, year of analysis, and path to save file}
}
\value{
df
}
\description{
This function functions as the mean function, only calculates the mean with
NAs removed. If data has NAs, there will no longer be an error.
}
