% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{demos}
\alias{demos}
\title{Demographic Tab}
\usage{
demos(df1, df2, df3, year, path = "", name = "ds_bl_")
}
\arguments{
\item{df1, df2, year, path, name}{Input df1 = original/consent df, df2 = panel data, year, path is default to current, name is set to "ds_bl_"}
}
\value{
df with demo vars cleaned up
}
\description{
This function cleans up the demographic variables
}
