% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{essay}
\alias{essay}
\title{Essay Tab}
\usage{
essay(df1, df2, path = "")
}
\arguments{
\item{df1, df2, path}{Input dataframes, path is default to current}
}
\value{
df with additional intervention condition variable
}
\description{
Step 8
}
\details{
This function cleans up the essay response.
}
