% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hist_fun}
\alias{hist_fun}
\title{Histogram}
\usage{
hist_fun(
  df,
  quant_vars,
  year,
  path = "output/EDA/plots/histograms/",
  na.rm = T,
  ...
)
}
\arguments{
\item{df, }{quant_vars,year,path Input dataframe, vector of variable names, year, and path is default to "output/EDA/plots/histograms/"}
}
\value{
histogram plots
}
\description{
This function creates an output of hist plots.
}
