% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{id_consent}
\alias{id_consent}
\title{Separate files by consent status}
\usage{
id_consent(df, lower_bound = 3, upper_bound = 95)
}
\arguments{
\item{df, lower_bound, upper_bound}{Input dataframe, lower_bound defaults to 3, upper_bound defaults to 95}
}
\value{
df with additional status variable
}
\description{
This function adds a consent status column based on the progress variable.
}
