% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{intervention_essay}
\alias{intervention_essay}
\title{Essay Tab}
\usage{
intervention_essay(df1, df2, year, path = "", name = "ie_bl_")
}
\arguments{
\item{df1, df2, year, path, name}{Input df1 = original/consent df, df2 = panel data, year, path is default to current, name default to "ie_bl_"}
}
\value{
df with additional intervention condition variable
}
\description{
This function cleans up the essay response.
}
