% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{itemify}
\alias{itemify}
\title{Items Tab}
\usage{
itemify(df, year, path = "")
}
\arguments{
\item{df1, year, path}{Input dataframes, year, and path is default to current}
}
\value{
df with additional intervention condition variable
}
\description{
Step 9
}
\details{
This function cleans up the response items.
}
