% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{manipulate_text}
\alias{manipulate_text}
\title{Manipulate Text Tab}
\usage{
manipulate_text(df)
}
\arguments{
\item{df1, df2, year, path, name}{Input df1 = original/consent df, df2 = panel data, year, path is default to current, name default to "ie_bl_"}
}
\value{
df with additional intervention condition variable
}
\description{
This function cleans up text responses.
}
