% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{mini_consent}
\alias{mini_consent}
\title{Separate files by mini consent status}
\usage{
mini_consent(df, year, path = "", name = "mc_bl_")
}
\arguments{
\item{df, year, path, }{consent_status Input dataframe, year of analysis, path, and consent_status can be specified "yes" or "no" but defaults to all}
}
\value{
df with additional consent variable
}
\description{
This function separates the dataframe to those who are open to be contacted with info.
}
