% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mean.R
\name{my_mean}
\alias{my_mean}
\title{Calculate a Mean}
\usage{
my_mean(x)
}
\arguments{
\item{x}{Input list or set of values}
}
\value{
mean of x
}
\description{
This function functions as the mean function, only calculates the mean with
NAs removed. If data has NAs, there will no longer be an error.
}
