% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{quality_control}
\alias{quality_control}
\title{Create quality control}
\usage{
quality_control(df, year, consent_df = NULL, path = "")
}
\arguments{
\item{df, year, consent_df, path}{Input dataframe, lower_bound defaults to 3, upper_bound defaults to 95}
}
\value{
df with additional consent variable
}
\description{
This function creates a quality control dataframe. User has the option to use
the full dataframe or filter based on the dataframe of students who have
consented to using their data for research. The consent_df argument is optional.
}
