% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{separate_consent}
\alias{separate_consent}
\title{Separate files by consent status}
\usage{
separate_consent(df, year, path = "")
}
\arguments{
\item{df, year, path}{Input dataframe, lower_bound defaults to 3, upper_bound defaults to 95}
}
\value{
df with additional consent variable
}
\description{
Step 5
}
\details{
This function separates the dataframe into three files: it writes two files
that separates groups by whether they consented or not. This function adds
a consent column to the original dataset and keeps all values in it.
}
