% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{update_reference}
\alias{update_reference}
\title{Baseline reference file}
\usage{
update_reference(df1, df2, year, period, path = "")
}
\arguments{
\item{df1, df2, year, path}{Input df1 = original key, df2 = new dataframe, year of analysis, and path to save new files}
}
\value{
original llave, modifed, and updated llave with added rows and IDs
}
\description{
Step 3
}
\details{
This function updates and creates new reference files using the original
key and new dataframe. The new key with added rows and IDs is produced, and
the original key file is modified with the year of update.
}
