% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar_fun}
\alias{bar_fun}
\title{Bar plot}
\usage{
bar_fun(
  df,
  cat_vars,
  year,
  path = "output/EDA/plots/barplots/",
  na.rm = T,
  ...
)
}
\arguments{
\item{df, cat_vars, year, path}{Input dataframe, vector of variable names, year, and path is default to "output/EDA/plots/barplots/"}
}
\value{
bar plots
}
\description{
This function creates an output of bar plots.
}
