% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(df, beg_row = 1, end_row = 2)
}
\arguments{
\item{df, beg_row, end_row}{Input dataframe, start row defaults to 1 and end row defaults to 2}
}
\value{
df
}
\description{
This function functions as the mean function, only calculates the mean with
NAs removed. If data has NAs, there will no longer be an error.
}
