% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{create_codebook}
\alias{create_codebook}
\title{Create codebook/ key file}
\usage{
create_codebook(df, year, path = "")
}
\arguments{
\item{df, year, path}{Input dataframe, year of analysis, and path to save file}
}
\value{
codebook dataframe
}
\description{
Step 1
}
\details{
This function creates a codebook with the first two rows from the survey,
which has the name and description of each variable. This function
should be run first.
}
