% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{deidentify}
\alias{deidentify}
\title{Create Deidentified file}
\usage{
deidentify(df, year, path = "")
}
\arguments{
\item{df, year, path}{Input dataframe, year of analysis, and path to save file}
}
\value{
df, also exported file
}
\description{
This function removed identifiers from the dataframe.
}
