% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{deidentify}
\alias{deidentify}
\title{Create De-identified file}
\usage{
deidentify(df, path = "", year)
}
\arguments{
\item{df, year, path}{Input file, year of analysis, and path}
}
\value{
df, also exported file
}
\description{
Step 4
}
\details{
This function removed identifiers from the dataframe.
}
