% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hist_fun}
\alias{hist_fun}
\title{Histogram}
\usage{
hist_fun(df, year, index, path = "output/EDA/plots/", na.rm = T, ...)
}
\arguments{
\item{df1, year, index, path}{Input dataframe, year, and path is default to current}
}
\value{
histogram plots
}
\description{
Histogram
}
\details{
This function creates an output of hist plots.
}
