% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{itemify}
\alias{itemify}
\title{Items Tab}
\usage{
itemify(df, year, path = "", name = "item_bl_")
}
\arguments{
\item{df1, year, path, name}{Input dataframe, year, path is default to current, name is set to "item_bl_"}
}
\value{
df with items cleaned up
}
\description{
This function cleans up the response items.
}
