% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R, R/tables.R
\name{quality_control}
\alias{quality_control}
\title{Identify intervention condition}
\usage{
quality_control(df, year, consent_df = NULL, path = "")

quality_control(df, year, consent_df = NULL, path = "")
}
\arguments{
\item{df, consent_df, path}{Input dataframe, consent_df defaults to NULL, path is default to current}

\item{df, year, consent_df, path}{Input dataframe, lower_bound defaults to 3, upper_bound defaults to 95}
}
\value{
df with additional intervention condition variable

df with additional consent variable
}
\description{
Step 7

This function creates a quality control dataframe. User has the option to use
the full dataframe or filter based on the dataframe of students who have
consented to using their data for research. The consent_df argument is optional.
}
\details{
This function adds mccm_correct column.
In other words, the intervention condition is specified.
}
