% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{quality_control}
\alias{quality_control}
\title{Identify intervention condition}
\usage{
quality_control(df, year, path = "", name = "qc_bl_")
}
\arguments{
\item{df, year, path, name}{Input dataframe, year, path is default to current, name default saves to "qc_bl_"}
}
\value{
df with additional intervention condition variable
}
\description{
This function adds mccm_correct column.
In other words, the intervention condition is specified.
}
