% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{separate_consent}
\alias{separate_consent}
\title{Separate files by consent status}
\usage{
separate_consent(df, year, path = "", consent_status = "")
}
\arguments{
\item{df, year, path, }{consent_status Input dataframe, year of analysis, path, and consent_status can be specified "yes" or "no" but defaults to all}
}
\value{
df with additional consent variable
}
\description{
Step 5
}
\details{
This function separates the dataframe into three files: it writes two files
that separates groups by whether they consented or not. This function adds
a consent column to the original dataset and keeps all values in it.
}
